refactor account module into new format
