--- apps/carbon/app/interfaces/Account/index.ts
@@ -1,3 +1,0 @@
-import useAccountSidebar from "./useAccountSidebar";
-
-export { useAccountSidebar };
