+++ apps/carbon/app/modules/account/interfaces/index.ts
@@ -1,0 +1,7 @@
+import useAccountSidebar from "./useAccountSidebar";
+
+export { useAccountSidebar };
+export * from "./Password";
+export * from "./Personal";
+export * from "./Profile";
+export * from "./UserAttributes";
