+++ apps/carbon/app/routes/x+/account+/$userId.delete.attribute.tsx
@@ -2,7 +2,7 @@ import type { ActionArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import { requirePermissions } from "~/services/auth";
 import { flash } from "~/services/session";
-import { deleteUserAttributeValue } from "~/services/account";
+import { deleteUserAttributeValue } from "~/modules/account";
 import { getAttribute } from "~/services/resources";
 import { getUserClaims } from "~/services/users";
 import { assertIsPost } from "~/utils/http";
