+++ apps/carbon/app/routes/x+/account+/password.tsx
@@ -3,8 +3,7 @@ import { json, redirect } from "@remix-run/router";
 import { validationError } from "remix-validated-form";
 import { PageTitle } from "~/components/Layout";
 import { getSupabaseServiceRole } from "~/lib/supabase";
-import { PasswordForm } from "~/interfaces/Account/Password";
-import { accountPasswordValidator } from "~/services/account";
+import { PasswordForm, accountPasswordValidator } from "~/modules/account";
 import { requirePermissions } from "~/services/auth";
 import { flash } from "~/services/session";
 import { assertIsPost } from "~/utils/http";
