+++ apps/carbon/app/routes/x+/account+/personal.tsx
@@ -4,9 +4,8 @@ import { redirect } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
 import { json } from "@remix-run/router";
 import { PageTitle, SectionTitle } from "~/components/Layout";
-import { UserAttributesForm } from "~/interfaces/Account/UserAttributes";
-import type { PrivateAttributes } from "~/interfaces/Account/types";
-import { getPrivateAttributes } from "~/services/account";
+import type { PrivateAttributes } from "~/modules/account";
+import { UserAttributesForm, getPrivateAttributes } from "~/modules/account";
 import { requirePermissions } from "~/services/auth";
 import { flash } from "~/services/session";
 import { error } from "~/utils/result";
